+++ src/full/Agda/TypeChecking/RecordPatterns.hs
@@ -7,10 +7,11 @@ module Agda.TypeChecking.RecordPatterns
   , translateCompiledClauses
   , translateSplitTree
   , recordPatternToProjections
+  , coinductiveRecordRHSsToCopatterns
   ) where
 
 import Control.Arrow          ( first, second )
-import Control.Monad          ( forM, unless, when, zipWithM )
+import Control.Monad          ( forM, join, unless, when, zipWithM )
 import Control.Monad.Fix      ( mfix )
 import Control.Monad.IO.Class ( MonadIO(..) )
 import Control.Monad.Reader   ( MonadReader(..), ReaderT(..), runReaderT )
@@ -38,11 +39,14 @@ import Agda.TypeChecking.Telescope
 import Agda.Interaction.Options
 
 import Agda.Utils.Either
+import Agda.Utils.Function
 import Agda.Utils.Functor
+import Agda.Utils.Monad
 import Agda.Utils.Permutation hiding (dropFrom)
-import Agda.Utils.Pretty (Pretty(..))
+import Agda.Utils.Pretty (Pretty(..), prettyShow)
 import qualified Agda.Utils.Pretty as P
 import Agda.Utils.Size
+import Agda.Utils.Tuple
 import Agda.Utils.Update (MonadChange, tellDirty)
 
 import Agda.Utils.Impossible
@@ -205,6 +209,78 @@ mergeCatchAll cc ca = maybe cc (mappend cc) ca
     _                   -> __IMPOSSIBLE__ -- this would mean non-determinism
 -}
 -}
+
+
+-- | Transform definitions returning coinductive record values to use copatterns instead.
+--   This allows termination-checking constructor-style coinduction.
+--
+--   For example:
+--
+--   @
+--     nats : Nat → Stream Nat
+--     nats n = n ∷ nats (1 + n)
+--   @
+--
+--   The clause is translated to:
+--
+--   @
+--     nats n .head = n
+--     nats n .tail = nats (1 + n)
+--   @
+--
+coinductiveRecordRHSsToCopatterns ::
+     forall m. (HasConstInfo m, PureTCM m)
+  => [Clause]
+  -> m [Clause]
+coinductiveRecordRHSsToCopatterns cls = do
+  reportSLn "tc.inline.con" 40 $ "enter coinductiveRecordRHSsToCopatterns with " ++ show (length cls) ++ " clauses"
+  flip concatMapM cls $ \case
+
+    -- RHS must be fully applied coinductive constructor/record expression.
+    cl@Clause{ namedClausePats = ps
+             , clauseBody      = Just v0@(Con con@(ConHead c _ _ind fs) _ci es)
+             , clauseType      = mt
+             }
+      | not (null fs)           -- at least one field
+      , length fs == length es  -- fully applied
+      , Just vs <- allApplyElims es
+
+          -- Only expand constructors labelled @{-# INLINE c #-}@.
+        -> ifNotM (inlineConstructor c) (return [cl]) {-else-} do
+
+          -- Iterate the translation for nested constructor rhss.
+          coinductiveRecordRHSsToCopatterns =<< do
+
+            -- Create one clause per projection.
+            forM (zip fs vs) $ \ (f, v) -> do
+
+              -- Get the type of the field.
+              let inst :: Type -> m (Maybe Type)
+                  inst t = fmap thd3 <$> projectTyped v0 t ProjSystem (unArg f)
+
+              let fuse :: Maybe (Arg (Maybe a)) -> Maybe (Arg a)
+                  fuse = join . fmap distributeF
+
+              mt' :: Maybe (Arg Type) <- fuse <$> traverse (traverse inst) mt
+
+              -- Make clause ... .f = v
+              return cl
+                { namedClausePats = ps ++ [ unnamed . ProjP ProjSystem <$> f ]
+                , clauseBody      = Just $ unArg v
+                , clauseType      = mt'
+                }
+
+    -- Otherwise: no change.
+    cl -> return [cl]
+
+  where
+    inlineConstructor :: QName -> m Bool
+    inlineConstructor c = getConstInfo c <&> theDef >>= \case
+      Constructor { conInline } -> do
+        reportSLn "tc.inline.con" 80 $
+          ("can" ++) $ applyUnless conInline ("not" ++) $ " inline constructor " ++ prettyShow c
+        return conInline
+      _ -> return False
 
 -- | Transform definitions returning record expressions to use copatterns
 --   instead. This prevents terms from blowing up when reduced.