+++ src/full/Agda/TypeChecking/Rules/Decl.hs
@@ -63,6 +63,7 @@ import Agda.TypeChecking.Rules.Display ( checkDisplayPragma )
 
 import Agda.Termination.TermCheck
 
+import Agda.Utils.Function ( applyUnless )
 import Agda.Utils.Functor
 import Agda.Utils.Lens
 import Agda.Utils.Maybe
@@ -786,8 +787,9 @@ checkPragma r p =
           def <- getConstInfo x
           case theDef def of
             Function{} -> markInline b x
-            _          -> typeError $ GenericError $ sINLINE ++ " directive only works on functions"
-              where sINLINE = if b then "INLINE" else "NOINLINE"
+            d@Constructor{ conSrcCon } | copatternMatchingAllowed conSrcCon
+              -> modifyGlobalDefinition x $ set lensTheDef d{ conInline = b }
+            _ -> typeError $ GenericError $ applyUnless b ("NO" ++) "INLINE directive only works on functions or constructors of records that allow copattern matching"
         A.OptionsPragma{} -> typeError $ GenericError $ "OPTIONS pragma only allowed at beginning of file, before top module declaration"
         A.DisplayPragma f ps e -> checkDisplayPragma f ps e
         A.EtaPragma r -> do