+++ src/full/Agda/Utils/Monad.hs
@@ -139,6 +139,9 @@ forMM_ = flip mapMM_
 
 -- Lists and monads -------------------------------------------------------
 
+concatMapM :: Monad m => (a -> m [b]) -> [a] -> m [b]
+concatMapM f xs = concat <$> Trav.mapM f xs
+
 -- | A monadic version of @'mapMaybe' :: (a -> Maybe b) -> [a] -> [b]@.
 mapMaybeM :: Monad m => (a -> m (Maybe b)) -> [a] -> m [b]
 mapMaybeM f xs = catMaybes <$> Trav.mapM f xs